Cleanup mcp_tools and general_tools
